third_party/WebKit/LayoutTests/platform/win/external/wpt/web-animations/interfaces/Animation/onfinish-expected.txt

This is a testharness.js-based test.
FAIL onfinish event is fired when the currentTime < 0 and the playbackRate < 0 assert_equals: event.timelineTime should equal to the animation timeline when finished promise is resolved expected (undefined) undefined but got (number) 309.2480003833771
FAIL onfinish event is fired when the currentTime > 0 and the playbackRate > 0 assert_equals: event.timelineTime should equal to the animation timeline when finished promise is resolved expected (undefined) undefined but got (number) 309.2480003833771
PASS onfinish event is fired when animation.finish() is called 
PASS onfinish event is not fired when paused 
PASS onfinish event is not fired when the playbackRate is zero 
PASS onfinish event is not fired when the animation falls out finished state immediately 
Harness: the test ran to completion.
